requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.4
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
Django==4.2
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
msgpack==1.0.5
pika==1.3.2
platformdirs==3.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pylint==2.17.3
pytz==2023.3
six==1.10.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
